+++ changelog.md
@@ -1,3 +1,10 @@
+# v1.0.13
+
+New Feature:
+
+* + InDevPointer
+* spi.setup() 支持4线SPI
+
 
 # v1.0.11 2026/5/11